[python-frontend] bytes data encoded in base64 to avoid escape notation during json dumping
