+++ src/util/encoding.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <vector>
+#include <cstdint>
+#include <string>
+
+std::vector<uint8_t> base64_decode(const std::string &encoded_string);
